style fixing for vacancies

+++ internal/network/handlers/vacancy_activity.go
@@ -34,7 +34,7 @@ func (vah *VacancyActivityHandler) GetAllVacancyApplies(c *gin.Context) {
 		return
 	}
 
-	applies, getAllErr := vah.vacancyActivityUseCase.GetAllVacancyApplies(id)
+	applies, getAllErr := vah.vacancyActivityUseCase.GetAllVacancyApplies(uint(id))
 	if getAllErr != nil {
 		_ = c.Error(getAllErr)
 		return
@@ -60,7 +60,7 @@ func (vah *VacancyActivityHandler) ApplyForVacancy(c *gin.Context) {
 		_ = c.Error(errorHandler.ErrBadRequest)
 		return
 	}
-	applyErr := vah.vacancyActivityUseCase.ApplyForVacancy(userId, id, &input)
+	applyErr := vah.vacancyActivityUseCase.ApplyForVacancy(userId, uint(id), &input)
 	if applyErr != nil {
 		_ = c.Error(applyErr)
 		return
@@ -78,7 +78,7 @@ func (vah *VacancyActivityHandler) GetAllUserApplies(c *gin.Context) {
 		return
 	}
 
-	applies, getErr := vah.vacancyActivityUseCase.GetAllUserApplies(userId)
+	applies, getErr := vah.vacancyActivityUseCase.GetAllUserApplies(uint(userId))
 	if getErr != nil {
 		_ = c.Error(getErr)
 		return
